+++ option-antd/src/data/EditData.js
@@ -144,10 +144,10 @@ class EditData extends BaseData {
       if (editdata.pagination) {
         if (!this.func.page) {
           if (!this.getData) {
-            this.printMsg('选择器存在分页器时需要定义page回调或者getData函数供分页时调用')
+            this.$exportMsg('选择器存在分页器时需要定义page回调或者getData函数供分页时调用')
           }
           this.func.page = (act, data) => {
-            this.loadData(true, this.option.search.value).then(res => {}, err => { this.printMsg('loadData失败！', 'error', { data: err }) })
+            this.loadData(true, this.option.search.value).then(res => {}, err => { this.$exportMsg('loadData失败！', 'error', { data: err }) })
           }
         }
         let paginationOption = editdata.pagination
@@ -227,7 +227,7 @@ class EditData extends BaseData {
               } else {
                 // 不强制加载
                 if (this.func.autoSearch('init')) {
-                  this.loadData(false, this.option.search.value).then(res => {}, err => { this.printMsg('loadData失败！', 'error', { data: err }) })
+                  this.loadData(false, this.option.search.value).then(res => {}, err => { this.$exportMsg('loadData失败！', 'error', { data: err }) })
                 }
               }
             }
@@ -282,7 +282,7 @@ class EditData extends BaseData {
           this.func.searchStart = (value) => {
             this.option.search.value = value || ''
             if (this.func.autoSearch('init')) {
-              this.loadData(true, this.option.search.value).then(res => {}, err => { this.printMsg('loadData失败！', 'error', { data: err }) })
+              this.loadData(true, this.option.search.value).then(res => {}, err => { this.$exportMsg('loadData失败！', 'error', { data: err }) })
             }
           }
         }
@@ -432,7 +432,7 @@ class EditData extends BaseData {
       this.option.fileUpload = editdata.option.fileUpload || false
       this.option.layout = editdata.option.layout === undefined ? 'auto' : editdata.option.layout
       if (this.option.upload && !this.option.fileUpload) {
-        this.printMsg('上传插件需要定义上传方法:fileUpload=>option')
+        this.$exportMsg('上传插件需要定义上传方法:fileUpload=>option')
       }
       if (this.option.multiple) {
         this.setValueToArray()
@@ -550,7 +550,7 @@ class EditData extends BaseData {
   _selfName () {
     return `(${this.constructor._name})`
   }
-  printMsg(info, type = 'error', option) {
+  $exportMsg(info, type = 'error', option) {
     _func.exportMsg(this._selfName() + ':' + info, type, option)
   }
   toString() {